+++ test/schema.test.ts
@@ -19,6 +19,93 @@ describe('Schema', () => {
 		expect(res.status).toBe(200)
 	})
 
+	it('validate params', async () => {
+		const app = new Elysia().get(
+			'/hi/:id/:name',
+			({ params: { id, name } }) => name,
+			{
+				schema: {
+					params: t.Object({
+						id: t.String(),
+						name: t.String()
+					})
+				}
+			}
+		)
+		const res = await app.handle(req('/hi/1/sucrose'))
+
+		expect(await res.text()).toBe('sucrose')
+		expect(res.status).toBe(200)
+	})
+
+	it('validate headers', async () => {
+		const app = new Elysia().post('/', () => 'welcome back', {
+			schema: {
+				headers: t.Object({
+					authorization: t.String()
+				})
+			}
+		})
+		const res = await app.handle(
+			new Request('/', {
+				method: 'post',
+				headers: {
+					authorization: 'Bearer 123',
+					// optional header should be allowed
+					'x-forwarded-ip': '127.0.0.1'
+				}
+			})
+		)
+
+		expect(await res.text()).toBe('welcome back')
+		expect(res.status).toBe(200)
+	})
+
+	it('validate body', async () => {
+		const app = new Elysia().post('/', ({ body }) => body, {
+			schema: {
+				body: t.Object({
+					username: t.String(),
+					password: t.String()
+				})
+			}
+		})
+
+		const body = JSON.stringify({
+			username: 'ceobe',
+			password: '12345678'
+		})
+
+		const res = await app.handle(
+			new Request('/', {
+				method: 'post',
+				body,
+				headers: {
+					'content-type': 'application/json'
+				}
+			})
+		)
+
+		expect(await res.text()).toBe(body)
+		expect(res.status).toBe(200)
+	})
+
+	it('validate response', async () => {
+		const app = new Elysia().get(
+			'/',
+			() => 'Mutsuki need correction 💢💢💢',
+			{
+				schema: {
+					response: t.String()
+				}
+			}
+		)
+		const res = await app.handle(req('/'))
+
+		expect(await res.text()).toBe('Mutsuki need correction 💢💢💢')
+		expect(res.status).toBe(200)
+	})
+
 	it('handle guard hook', async () => {
 		const app = new Elysia().guard(
 			{